Reorganize jetstream module
Merges all jetstream prefixed modules into a single module with
sub-modules.
