+++ src/jetstream/mod.rs
@@ -113,9 +113,11 @@ use serde_repr::{Deserialize_repr, Serialize_repr};
 
 const ORDERED_IDLE_HEARTBEAT: Duration = Duration::from_nanos(5_000_000_000);
 
-// TODO re-organize this into a jetstream directory
-pub use crate::jetstream_push_subscription::PushSubscription;
-pub use crate::jetstream_types::*;
+mod push_subscription;
+mod types;
+
+pub use push_subscription::PushSubscription;
+pub use types::*;
 
 use crate::{
     header::{self, HeaderMap},
